requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appdirs=1.4.4=pyh9f0ad1d_0
archspec=0.2.1=pyhd3eb1b0_0
attrs=23.1.0=pyh71513ae_1
automat=22.10.0=pyhd8ed1ab_0
bcrypt=3.2.0=py310hca72f7f_1
beautifulsoup4=4.12.2=pypi_0
boltons=23.0.0=py310hecd8cb5_0
brotli-python=1.0.9=py310he9d5cce_7
bzip2=1.0.8=h1de35cc_0
c-ares=1.19.1=h6c40b1e_0
ca-certificates=2023.08.22=hecd8cb5_0
certifi=2023.11.17=py310hecd8cb5_0
cffi=1.16.0=py310h6c40b1e_0
cfscrape=2.1.1=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudscraper=1.2.71=pypi_0
conda=23.11.0=py310hecd8cb5_0
conda-libmamba-solver=23.11.1=py310hecd8cb5_0
conda-package-handling=2.2.0=py310hecd8cb5_0
conda-package-streaming=0.9.0=py310hecd8cb5_0
constantly=15.1.0=py_0
cryptography=41.0.3=py310h30e54ef_0
cssselect=1.2.0=pyhd8ed1ab_0
distro=1.5.0=pyhd3eb1b0_1
filelock=3.13.1=pyhd8ed1ab_0
fmt=9.1.0=ha357a0b_0
hyperlink=21.0.0=pyhd3deb0d_0
icu=73.1=hcec6c5f_0
idna=3.4=py310hecd8cb5_0
incremental=22.10.0=pyhd8ed1ab_0
itemadapter=0.8.0=pyhd8ed1ab_0
itemloaders=1.1.0=pyhd8ed1ab_0
jmespath=1.0.1=pyhd8ed1ab_0
jsonpatch=1.32=pyhd3eb1b0_0
jsonpointer=2.1=pyhd3eb1b0_0
krb5=1.20.1=h428f121_1
libarchive=3.6.2=h29ab7a1_2
libcurl=8.4.0=hf20ceda_0
libcxx=14.0.6=h9765a3e_0
libedit=3.1.20221030=h6c40b1e_0
libev=4.33=h9ed2024_1
libffi=3.4.4=hecd8cb5_0
libiconv=1.16=hca72f7f_2
libmamba=1.5.3=h63cd6dc_0
libmambapy=1.5.3=py310h8c3233a_0
libnghttp2=1.57.0=h9beae6a_0
libsolv=0.7.24=hfff2838_0
libssh2=1.10.0=h04015c4_2
libxml2=2.10.4=h1bd7e62_1
libxslt=1.1.37=h6c40b1e_1
lxml=4.9.3=py310h946e0e5_0
lz4-c=1.9.4=hcec6c5f_0
menuinst=2.0.0=py310hecd8cb5_0
ncurses=6.4=hcec6c5f_0
numpy=1.26.2=pypi_0
openssl=3.0.12=hca72f7f_0
packaging=23.1=py310hecd8cb5_0
pandas=2.1.4=pypi_0
parsel=1.8.1=pyhd8ed1ab_0
pcre2=10.42=h9b97e30_0
pip=23.3=py310hecd8cb5_0
platformdirs=3.10.0=py310hecd8cb5_0
pluggy=1.0.0=py310hecd8cb5_1
protego=0.3.0=pyhd8ed1ab_0
pyasn1=0.5.1=pyhd8ed1ab_0
pyasn1-modules=0.3.0=pyhd8ed1ab_0
pybind11-abi=4=hd3eb1b0_1
pycosat=0.6.6=py310h6c40b1e_0
pycparser=2.21=pyhd3eb1b0_0
pydispatcher=2.0.5=py_1
pyopenssl=23.2.0=py310hecd8cb5_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py310hecd8cb5_0
python=3.10.13=h5ee71fb_0
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pypi_0
pytz=2023.3.post1=pypi_0
queuelib=1.6.2=pyhd8ed1ab_0
readline=8.2=hca72f7f_0
reproc=14.2.4=he9d5cce_1
reproc-cpp=14.2.4=he9d5cce_1
requests=2.31.0=py310hecd8cb5_0
requests-file=1.5.1=pyh9f0ad1d_0
requests-toolbelt=1.0.0=pypi_0
ruamel.yaml=0.17.21=py310hca72f7f_0
ruamel.yaml.clib=0.2.6=py310hca72f7f_1
scrapy=2.8.0=py310hecd8cb5_0
scrapy-cloudflare-middleware=0.0.1=pypi_0
service-identity=21.1.0=pyhd8ed1ab_0
service_identity=21.1.0=hd8ed1ab_0
setuptools=68.0.0=py310hecd8cb5_0
six=1.16.0=pyh6c4a22f_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h6c40b1e_0
tk=8.6.12=h5d9f67b_0
tldextract=5.1.1=pyhd8ed1ab_0
tqdm=4.65.0=py310h20db666_0
truststore=0.8.0=py310hecd8cb5_0
typing-extensions=4.9.0=hd8ed1ab_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023.3=pypi_0
urllib3=1.26.18=py310hecd8cb5_0
w3lib=2.1.2=pyhd8ed1ab_0
wheel=0.41.2=py310hecd8cb5_0
xz=5.4.2=h6c40b1e_0
yaml-cpp=0.8.0=hcec6c5f_0
zlib=1.2.13=h4dc903c_0
zope=1.0=py310hecd8cb5_1
zope.interface=5.4.0=py310hca72f7f_0
zstandard=0.19.0=py310h6c40b1e_0
zstd=1.5.5=hc035e20_0

schedule~=1.2.1
scrapy~=2.8.0
python-dotenv~=0.21.0
itemadapter~=0.3.0
requests~=2.31.0
twisted~=22.10.0